Move code related to Azure Service Connection to where the env vars are gathered for the Pulumi command.
